Fixed: Fix exit code handling
strtol() returns 0 on conversion failure and errno is not set, which resulted in wrong exit codes passed by server to be assumed to be exit code 0.

Moreover, shell exit codes are limited to 0-255 and that was not being checked. Out of bound exit codes would return with exit code `44` `Channel number out of range`.