Fixed: Fix argument parsing and convert termux-am binary to termux-am-socket and introduce termux-am wrapper script
Arguments were being passed to server by surrounding them with double quotes, which obviously wouldn't work if arguments themselves contained double quotes and would result in broken commands and wrong data.

bash printf built-in provides `%q` format specification which can be used to convert argument arrays to a single string that can be reused as shell input. The `ArgumentTokenizer` should be able to handle this string to convert it back to arguments. The c printf does not support `%q` and porting bash functionality would require importing a lot of code. So the new termux-am script will convert arguments to a string and pass it to the termux-am-socket binary to be sent to the server.

The termux-am also provides a help screen to help user understand how it works and provides `--am-help` flag to show am command help instead which is returned by termux-am-library.